Extract tx type enum
